xpenology: Boot param for number of XEN PV HDDs.
To resolve a crash where the offset used for internal XEN devices
clashes with non XEN virtualised hard drives. This boot param
can be used to determine which min offset the other internal drives
should use. The crash is due to the way PV drives are handled
using xen block drivers and not the traditonal scsi driver.

i.e xen (sda,sdb) = 2

Ideally this should be automated and not manually set by the user.